requirements.txt
paramiko~=2.12.0
click~=8.1.3
PyYAML~=6.0